[feat] 코인 생성
코인생성
