+++ Assets/Data/WeaponData/Violin.asset
@@ -20,7 +20,7 @@ MonoBehaviour:
   icon: {fileID: 21300000, guid: 666b737749017194ea5fc01a3868ac57, type: 3}
   Price: 0
   Type: 3
-  AttackDamage: 3
+  AttackDamage: 1
   AttackSpeed: 0
   WeaponOverride: {fileID: 0}
   AttackRange: 0
