+++ Assets/Student/JHT/Anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 92a48090d3372124fb98d9ed69611416
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
