+++ Assets/Student/JHT/Anim/CoinAnim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 290ae97add2138a418fe118313024b54
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
