+++ Assets/Student/JHT/Anim/CoinAnim.controller.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 83e1551b46fe0414e96b5ce56dde48ac
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 9100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
