+++ Assets/Student/JHT/Prefabs/Coin.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 2d020f594b37ceb4bbd71f997dd068db
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
